add back the major version check

+++ VMEncryption/main/patch/redhatPatching.py
@@ -143,31 +143,32 @@ class redhatPatching(AbstractPatching):
             self.command_executor.Execute("pip install adal")
 
     def update_prereq(self):
-        dracut_repack_needed = False
+        if (self.distro_info[1].startswith('7.')):
+            dracut_repack_needed = False
 
-        if os.path.exists("/lib/dracut/modules.d/91lvm/"):
-            # If 90lvm already exists 91lvm will cause problems, so remove it.
-            if os.path.exists("/lib/dracut/modules.d/90lvm/"):
-                shutil.rmtree("/lib/dracut/modules.d/91lvm/")
-            else:
-                os.rename("/lib/dracut/modules.d/91lvm/","/lib/dracut/modules.d/90lvm/")
-            dracut_repack_needed = True
+            if os.path.exists("/lib/dracut/modules.d/91lvm/"):
+                # If 90lvm already exists 91lvm will cause problems, so remove it.
+                if os.path.exists("/lib/dracut/modules.d/90lvm/"):
+                    shutil.rmtree("/lib/dracut/modules.d/91lvm/")
+                else:
+                    os.rename("/lib/dracut/modules.d/91lvm/","/lib/dracut/modules.d/90lvm/")
+                dracut_repack_needed = True
 
-        if redhatPatching.is_old_patching_system():
-            redhatPatching.remove_old_patching_system(self.logger, self.command_executor)
-            dracut_repack_needed = True
+            if redhatPatching.is_old_patching_system():
+                redhatPatching.remove_old_patching_system(self.logger, self.command_executor)
+                dracut_repack_needed = True
 
-        if os.path.exists("/lib/dracut/modules.d/91ade/"):
-            shutil.rmtree("/lib/dracut/modules.d/91ade/")
-            dracut_repack_needed = True
+            if os.path.exists("/lib/dracut/modules.d/91ade/"):
+                shutil.rmtree("/lib/dracut/modules.d/91ade/")
+                dracut_repack_needed = True
 
-        if os.path.exists("/dev/mapper/osencrypt"):
-            #TODO: only do this if needed (if code and existing module are different)
-            redhatPatching.add_91_ade_dracut_module(self.command_executor)
-            dracut_repack_needed = True
+            if os.path.exists("/dev/mapper/osencrypt"):
+                #TODO: only do this if needed (if code and existing module are different)
+                redhatPatching.add_91_ade_dracut_module(self.command_executor)
+                dracut_repack_needed = True
 
-        if dracut_repack_needed:
-            self.command_executor.ExecuteInBash("/usr/sbin/dracut -I ntfs-3g -f -v --kver `grubby --default-kernel | sed 's|/boot/vmlinuz-||g'`", True)
+            if dracut_repack_needed:
+                self.command_executor.ExecuteInBash("/usr/sbin/dracut -I ntfs-3g -f -v --kver `grubby --default-kernel | sed 's|/boot/vmlinuz-||g'`", True)
 
     @staticmethod
     def is_old_patching_system():
